revert to the version with best speed

+++ src/lib.rs
@@ -1,7 +1,7 @@
 #[macro_export]
 macro_rules! assert_delta {
     ($x:expr, $y:expr, $delta:expr) => {
-        if (f32::abs($x - $y) > $delta) {
+        if (f64::abs($x - $y) > $delta) {
             panic!("{} != {}", $x, $y);
         }
     };
